Update tests and write new ones for separation location
