+++ src/applications/disability-benefits/all-claims/tests/property-names.unit.spec.js
@@ -19,6 +19,7 @@ const ignoreList = {
     'militaryHistory',
     'militaryHistory',
     'reservesNationalGuardService',
+    'separationLocation',
   ],
   unemployability: [
     'medicalCare',
